+++ conda_forge_tick/migrators/__init__.py
@@ -11,3 +11,4 @@ from .core import (
 from .migration_yaml import MigrationYaml
 from .arch import ArchRebuild
 from .pip_check import PipCheckMigrator
+from .version import Version
